+++ tests/config_test.py
@@ -44,6 +44,7 @@ test_rule = {'es_host': 'test_host',
 test_args = mock.Mock()
 test_args.config = 'test_config'
 test_args.rule = None
+test_args.debug = False
 
 
 def test_import_rules():
@@ -227,6 +228,20 @@ def test_load_ssl_env_true():
                 rules = load_rules(test_args)
 
                 assert rules['use_ssl'] is True
+
+
+def test_load_disabled_rules():
+    test_rule_copy = copy.deepcopy(test_rule)
+    test_rule_copy['is_enabled'] = False
+    test_config_copy = copy.deepcopy(test_config)
+    with mock.patch('elastalert.config.yaml_loader') as mock_open:
+        mock_open.side_effect = [test_config_copy, test_rule_copy]
+
+        with mock.patch('os.listdir') as mock_ls:
+            mock_ls.return_value = ['testrule.yaml']
+            rules = load_rules(test_args)
+            # The rule is not loaded for it has "is_enabled=False"
+            assert len(rules['rules']) == 0
 
 
 def test_compound_query_key():